+++ tester/Makefile
@@ -6,7 +6,7 @@
 #    By: gsmereka <gsmereka@student.42.fr>          +#+  +:+       +#+         #
 #                                                 +#+#+#+#+#+   +#+            #
 #    Created: 2022/12/19 18:51:25 by gsmereka          #+#    #+#              #
-#    Updated: 2023/01/23 22:40:54 by gsmereka         ###   ########.fr        #
+#    Updated: 2023/01/24 12:52:40 by gsmereka         ###   ########.fr        #
 #                                                                              #
 # **************************************************************************** #
 
@@ -20,7 +20,7 @@ OBJ_DIR			=	./obj
 
 SRC_DIR			=	src
 
-SRC				=	main.c initialize.c finalize.c redirect.c
+SRC				=	main.c initialize.c finalize.c redirect.c  test_loop.c
 
 OBJ				=	$(addprefix $(OBJ_DIR)/, $(SRC:.c=.o))
