+++ tester/tests/user_outputs/test_0
@@ -1,1 +1,1 @@
-teste_1
+test - 0
