+++ tester/tests/user_outputs/test_1
@@ -1,1 +1,1 @@
-teste_1
+test - 1
