+++ tester/tests/user_outputs/test_2
@@ -1,1 +1,1 @@
-teste_1
+test - 2
